requirements.txt
flask==2.3.3
alpaca-py==0.8.0
pandas==1.5.3
numpy==1.24.3
statsmodels==0.14.0
scipy==1.10.1
python-dotenv==1.0.0
websocket-client==1.7.0
yfinance==0.2.36
beautifulsoup4==4.12.2
scikit-learn==1.3.0
requests==2.31.0 